test(e2e): fold tab-overflow handling into shared openRequestPaneTab
Move the narrow-pane '>>' overflow fallback from datatype-vars.spec into the
shared openRequestPaneTab helper so every spec reaches overflowed tabs; the
direct-click path is unchanged for callers whose tab is visible.

Addresses PR review feedback.
