+++ tests/e2e/utils/page/actions.ts
@@ -917,11 +917,16 @@ export async function pasteIntoCollection(sidebar: Frame, collectionName: string
  * @param tabName - Visible label of the tab to select
  */
 export async function openRequestPaneTab(editor: Frame, tabName: string): Promise<void> {
-  await editor
-    .locator('[role="tab"]')
-    .filter({ hasText: tabName })
-    .first()
-    .click();
+  // Request-pane tabs collapse into a ">>" overflow menu when the pane is narrow, so click the tab
+  // directly when it's visible, otherwise reach it via the overflow menu.
+  const directTab = editor.locator('[role="tab"]').filter({ hasText: tabName }).first();
+  if (await directTab.isVisible().catch(() => false)) {
+    await directTab.click();
+    return;
+  }
+  await editor.locator('.more-tabs').click();
+  await editor.locator(`[data-testid="menu-dropdown-${tabName.toLowerCase()}"]`).click();
+  await expect(directTab).toBeVisible({ timeout: 15_000 });
 }
 
 /**
